code/requirements.txt
Cython==3.0.0a8
sentence-transformers==1.2.0
numpy
scipy
pandas
matplotlib
clickhouse-driver==0.2.0
tqdm==4.38.0
joblib==0.14.0
# we install a non offfical faiss using pip install faiss-gpu==1.6.1, since its official version requires conda
faiss-gpu==1.6.5
nni==1.8
#git+https://github.com/YaYaB/shuffle-big-file
#MatchZoo==2.2.0
#Keras==2.3.0
#tensorflow-gpu==1.15.2
seaborn==0.11.0
rankereval==0.1.1
git+git://github.com/AmenRa/rank_eval@eafd6d1f9cf9f0b9b9aa5eaf6ec8acd02f359133#egg=rank_eval
# add trec_eval
pytrec_eval==0.5
git+git://github.com/epfml/sent2vec@f00a1b67f4330e5be99e7cc31ac28df94deed9ac#egg=sent2vec
git+git://github.com/carmanzhang/sentence2vec@cf308b2000f4af7180e294508c3c0a9044244896#egg=sentence2vec
fairseq==0.10.2
fastBPE==0.1.0
subword-nmt==0.3.7
sacremoses==0.0.43
pytorch-skipthoughts==0.4.4
# pip3 install https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz
spacy==2.2.0
snowballstemmer==2.1.0